+++ database/migrations/2025_04_05_123400_create_notifications_table.php
@@ -13,21 +13,23 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('notifications', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->text('description')->nullable();
-            $table->foreignId('sender_id')->nullable()->constrained('users')->onDelete('set null');
-            $table->foreignId('receiver_id')->nullable()->constrained('users')->onDelete('cascade');
-            $table->string('receiver_type')->nullable()->comment('user, group, role');
-            $table->foreignId('group_id')->nullable()->constrained('groups')->onDelete('cascade');
-            $table->string('role')->nullable()->comment('admin, teacher, student');
-            $table->timestamp('read_at')->nullable();
-            $table->string('notification_type')->default('general')->comment('general, academic, announcement, exam');
-            $table->string('url')->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('notifications')) {
+            Schema::create('notifications', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->text('description')->nullable();
+                $table->foreignId('sender_id')->nullable()->constrained('users')->onDelete('set null');
+                $table->foreignId('receiver_id')->nullable()->constrained('users')->onDelete('cascade');
+                $table->string('receiver_type')->nullable()->comment('user, group, role');
+                $table->foreignId('group_id')->nullable()->constrained('groups')->onDelete('cascade');
+                $table->string('role')->nullable()->comment('admin, teacher, student');
+                $table->timestamp('read_at')->nullable();
+                $table->string('notification_type')->default('general')->comment('general, academic, announcement, exam');
+                $table->string('url')->nullable();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
